requirements.txt
Pillow==2.4.0
numpy==1.8.1
